backend/services/azure_vision_service.py
"""
Azure Computer Vision service for OCR and image analysis.
Provides advanced text extraction from prescription images.
"""
import io
import time
import sys
import logging
import traceback
from msrest.exceptions import ClientRequestError, HttpOperationError
from azure.cognitiveservices.vision.computervision import ComputerVisionClient
from azure.cognitiveservices.vision.computervision.models import OperationStatusCodes
from msrest.authentication import CognitiveServicesCredentials
import random
from PIL import Image

# Import application configuration
sys.path.append('..')
from config import AZURE_VISION_KEY, AZURE_VISION_ENDPOINT, USE_AZURE_OCR

# Configure logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class AzureVisionService:
    def __init__(self):
        """
        Initialize the Azure Computer Vision client with credentials.
        """
        self.is_available = False
        
        # Skip Azure initialization if simulation mode is enabled
        if not USE_AZURE_OCR:
            logger.info("Azure OCR is disabled in config. Using simulation mode.")
            return
            
        # Try to initialize Azure Computer Vision client
        try:
            logger.info(f"Initializing Azure Vision with endpoint: {AZURE_VISION_ENDPOINT}")
            # Sanitize key for logging (show only first 5 chars)
            safe_key = AZURE_VISION_KEY[:5] + "..." if AZURE_VISION_KEY else "None"
            logger.info(f"Using key (first 5 chars): {safe_key}")
            
            if not AZURE_VISION_KEY or AZURE_VISION_KEY == "your_azure_vision_key_here":
                logger.error("❌ Invalid Azure key detected - using placeholder value")
                return
            
            if not AZURE_VISION_ENDPOINT or "your-resource-name" in AZURE_VISION_ENDPOINT:
                logger.error("❌ Invalid Azure endpoint detected - using placeholder value")
                return
            
            self.client = ComputerVisionClient(
                endpoint=AZURE_VISION_ENDPOINT,
                credentials=CognitiveServicesCredentials(AZURE_VISION_KEY)
            )
            
            # Validate connection by making a simple API call
            logger.info("Testing Azure connection...")
            try:
                # Simply get client properties - lightweight API call
                self.client.api_version
                self.is_available = True
                logger.info("✅ Azure Vision service initialized and connection verified!")
            except (ClientRequestError, HttpOperationError) as e:
                logger.error(f"❌ Failed to connect to Azure service: {str(e)}")
                return
                
        except Exception as e:
            logger.error(f"❌ Azure Vision service initialization error: {str(e)}")
            logger.error(traceback.format_exc())
            logger.warning("Continuing with simulation mode.")
            
    def extract_text(self, image_bytes):
        """
        Extract text from prescription image using Azure Computer Vision.
        
        Args:
            image_bytes: Image data as bytes
            
        Returns:
            List of dictionaries containing extracted text and metadata
        """
        if not self.is_available:
            logger.warning("Azure Vision service is not available. Using fallback method.")
            return self._fallback_extract_text(image_bytes)
            
        try:
            logger.info("Extracting text using Azure OCR...")
            # Convert bytes to stream
            image_stream = io.BytesIO(image_bytes)
            
            # Call the Read API with the image stream
            read_response = self.client.read_in_stream(image_stream, raw=True)
            
            # Get the operation location from the response headers
            operation_location = read_response.headers["Operation-Location"]
            
            # Extract the operation ID from the operation location
            operation_id = operation_location.split("/")[-1]
            
            # Wait for the operation to complete
            max_retries = 10
            polling_interval = 1  # seconds
            result = None
            
            for i in range(max_retries):
                try:
                    result = self.client.get_read_result(operation_id)
                    if result.status not in [OperationStatusCodes.running, OperationStatusCodes.not_started]:
                        break
                    time.sleep(polling_interval)
                    logger.info(f"Polling Azure OCR result... attempt {i+1}/{max_retries}")
                except Exception as e:
                    logger.error(f"Error polling OCR result: {str(e)}")
                    time.sleep(polling_interval)
            
            # Process the results if the operation was successful
            extracted_text = []
            if result and result.status == OperationStatusCodes.succeeded:
                for read_result in result.analyze_result.read_results:
                    for line in read_result.lines:
                        # Format similar to our original OCR service for compatibility
                        text_data = {
                            "text": line.text,
                            "confidence": sum([word.confidence for word in line.words]) / len(line.words) if line.words else 0.5,
                            "box": [[p.x, p.y] for p in line.bounding_box] if hasattr(line, 'bounding_box') and hasattr(line.bounding_box[0], 'x') else self._format_bounding_box(line.bounding_box) if hasattr(line, 'bounding_box') else [[0, 0], [1, 0], [1, 1], [0, 1]]
                        }
                        extracted_text.append(text_data)
                
                logger.info(f"✅ Azure OCR completed successfully. Extracted {len(extracted_text)} text lines.")
                
                if len(extracted_text) > 0:
                    logger.info(f"Sample text: {extracted_text[0]['text']}")
                
                return extracted_text
            else:
                status = result.status if result else "Unknown"
                logger.error(f"❌ Azure OCR failed with status: {status}")
                return self._fallback_extract_text(image_bytes)
            
        except Exception as e:
            logger.error(f"❌ Azure OCR error: {str(e)}")
            logger.error(traceback.format_exc())
            return self._fallback_extract_text(image_bytes)
            
    def _fallback_extract_text(self, image_bytes):
        """
        Fallback method when Azure OCR fails or is unavailable.
        Uses a basic image processing approach to extract text structures.
        """
        import random
        import io
        import traceback
        from PIL import Image
        
        # Try to process the image to simulate some basic text extraction
        try:
            # Open image from bytes
            image = Image.open(io.BytesIO(image_bytes))
            width, height = image.size
            logger.info(f"Processing image: {width}x{height} pixels")
            
            # For fallback, divide the image into sections that might contain text
            # This simulates finding text blocks without actual OCR
            sections = []
            
            # Top section (patient info)
            sections.append((0, 0, width, height * 0.2))
            
            # Middle sections (potential prescription items)
            row_height = height * 0.15
            for i in range(1, 5):  # 4 potential rows of prescriptions
                y_start = height * 0.2 + (i-1) * row_height
                sections.append((0, y_start, width, y_start + row_height))
            
            # Bottom section (doctor info)
            sections.append((0, height * 0.8, width, height))
            
            # Create generic text blocks without specific medication content
            extracted_text = []
            for i, section in enumerate(sections):
                # Generate a more generic text label for each section
                if i == 0:
                    text = "Patient Information Section"
                elif i == len(sections) - 1:
                    text = "Doctor Information Section"
                else:
                    text = f"Prescription Item {i}"
                
                # Add text block with section coordinates
                text_data = {
                    "text": text,
                    "confidence": random.uniform(0.7, 0.9),
                    "box": [[section[0], section[1]], 
                            [section[2], section[1]], 
                            [section[2], section[3]], 
                            [section[0], section[3]]]
                }
                extracted_text.append(text_data)
            
            logger.info(f"Fallback OCR created {len(extracted_text)} generic text sections")
            return extracted_text
                
        except Exception as e:
            logger.error(f"Error in fallback OCR: {str(e)}")
            logger.error(traceback.format_exc())
            
            # Return minimal generic result if everything fails
            return [{
                "text": "Unable to process prescription image",
                "confidence": 0.5,
                "box": [[0, 0], [100, 0], [100, 20], [0, 20]]
            }]

    def _format_bounding_box(self, bounding_box):
        """
        Format the bounding box values correctly.
        The Azure OCR API can return bounding boxes in two formats:
        1. A list of Point objects with x,y properties
        2. A list of float values representing [x1,y1,x2,y2,x3,y3,x4,y4]
        
        This function handles both formats and returns a list of [x,y] coordinates.
        """
        try:
            # If bounding_box is None or empty, return default box
            if not bounding_box:
                return [[0, 0], [1, 0], [1, 1], [0, 1]]
            
            # Check if the bounding box is a list of floats [x1,y1,x2,y2,...]
            if isinstance(bounding_box[0], (int, float)):
                # The format is [x1,y1,x2,y2,x3,y3,x4,y4]
                if len(bounding_box) == 8:
                    return [
                        [bounding_box[0], bounding_box[1]],  # top-left
                        [bounding_box[2], bounding_box[3]],  # top-right
                        [bounding_box[4], bounding_box[5]],  # bottom-right
                        [bounding_box[6], bounding_box[7]]   # bottom-left
                    ]
                # The format is [x1,y1,w,h]
                elif len(bounding_box) == 4:
                    x, y, w, h = bounding_box
                    return [
                        [x, y],           # top-left
                        [x + w, y],       # top-right
                        [x + w, y + h],   # bottom-right
                        [x, y + h]        # bottom-left
                    ]
            
            # Default format for any other case
            return [[0, 0], [1, 0], [1, 1], [0, 1]]
        except Exception as e:
            logger.error(f"Error formatting bounding box: {e}")
            # Return default box on error
            return [[0, 0], [1, 0], [1, 1], [0, 1]]
