requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.0.76
numpy==1.25.2
httpx==0.24.1
pydantic==2.4.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-core==1.29.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1 